requirements.txt
Flask==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
flake8==7.1.0